refactor: extract DiagnosticsEvent.swift from DiagnosticsTypes.swift
Splits the `DiagnosticsEvent` event-name namespace + the paired
DiagnosticsEventName sugar extension out of DiagnosticsTypes.swift.

Byte-identical move of lines 54-92. New file needs only Foundation.

DiagnosticsTypes.swift: 92 → 53 lines. Tests: 667.
